+++ src/app/components/navbar/navbarController.js
@@ -149,28 +149,29 @@ angular.module('liveopsConfigPanel')
       };
 
       $scope.logout = function() {
-        var monitoredInteraction = CxEngage.session.getMonitoredInteraction();
-        if (monitoredInteraction !== null) {
-          var confirmedToLogout = confirm($translate.instant('interactionMonitoring.confirmEnd.logout'));
-          if (confirmedToLogout) {
+        // TODO: When old silent monitoring is removed we can uncomment out the below
+        // var monitoredInteraction = CxEngage.session.getMonitoredInteraction();
+        // if (monitoredInteraction !== null) {
+        //   var confirmedToLogout = confirm($translate.instant('interactionMonitoring.confirmEnd.logout'));
+        //   if (confirmedToLogout) {
             
-              CxEngage.interactions.voice.resourceRemove({interactionId: monitoredInteraction}, function() {
-                setTimeout(function() {
-                  AuthService.logout();
-                  $state.transitionTo('login');
-                  $rootScope.$broadcast('logout');
-                  // Reload ensure no saved state for the next session
-                  location.reload();
-                },1000);
-              });
+        //       CxEngage.interactions.voice.resourceRemove({interactionId: monitoredInteraction}, function() {
+        //         setTimeout(function() {
+        //           AuthService.logout();
+        //           $state.transitionTo('login');
+        //           $rootScope.$broadcast('logout');
+        //           // Reload ensure no saved state for the next session
+        //           location.reload();
+        //         },1000);
+        //       });
             
-          }
-        } else {
+        //   }
+        // } else {
           AuthService.logout();
           $state.transitionTo('login');
           $rootScope.$broadcast('logout');
           location.reload();
-        }
+        // }
       };
 
       $scope.userDropdownItems = [{
@@ -505,7 +506,7 @@ angular.module('liveopsConfigPanel')
 
           items.push({
             label: $translate.instant('navbar.reports.silentMonitoring.title'),
-            stateLink: 'content.reporting.interactionMonitoring',
+            stateLink: 'content.reporting.silentMonitoring',
             id: 'silent-monitoring-link',
             order: 5
           });