+++ src/app/shared/directives/sdkListener/sdkListenerDirective.js
@@ -16,22 +16,24 @@ angular.module('liveopsConfigPanel')
                 if (event.data.module === 'subscribe') {
                   var subscribedTenant = CxEngage.session.getActiveTenantId();
                   window.cxSubscriptions[event.data.command + subscribedTenant] = CxEngage.subscribe(event.data.command, function(error, topic, response) {
-                      if (location.hash.indexOf('#/reporting/interactionMonitoring') < 0) {
-                        CxEngage.unsubscribe(window.cxSubscriptions[event.data.command]);
-                        CxEngage.reporting.removeStatSubscription({ statId: 'interactions-in-conversation-list' });
-                      } else {
-                        try {
-                          event.source.postMessage({
-                            subscription: {
-                              error: error,
-                              topic: topic,
-                              response: response
-                            }
-                          }, '*');
-                        } catch (error) {
-                          console.warn('Cannot find original requestor iframe for subscription: ' ,event.data.command + 'for tenant id ' + subscribedTenant)
-                          CxEngage.unsubscribe(window.cxSubscriptions[event.data.command + subscribedTenant]);
+                      if (location.hash.indexOf('#/reporting/interactionMonitoring?alpha') < 0) {
+                        if(event.data.command === 'cxengage/reporting/batch-response') {
+                          CxEngage.unsubscribe(window.cxSubscriptions[event.data.command]);
                         }
+                        CxEngage.reporting.removeStatSubscription({ statId: 'interactions-in-conversation-list' });
+                      } 
+                   
+                      try {
+                        event.source.postMessage({
+                          subscription: {
+                            error: error,
+                            topic: topic,
+                            response: response
+                          }
+                        }, '*');
+                      } catch (error) {
+                        console.warn('Cannot find original requestor iframe for subscription: ' ,event.data.command + 'for tenant id ' + subscribedTenant)
+                        CxEngage.unsubscribe(window.cxSubscriptions[event.data.command + subscribedTenant]);
                       }
                     }
                   );
@@ -92,7 +94,7 @@ angular.module('liveopsConfigPanel')
                   event.source.postMessage({
                     error: null,
                     topic: ['updateLocalStorage'],
-                    response: JSON.parse(window.localStorage.getItem('LIVEOPS-PREFERENCE-KEY')).tenant
+                    response: {tenant: JSON.parse(window.localStorage.getItem('LIVEOPS-PREFERENCE-KEY')).tenant, agentId: CxEngage.session.getActiveUserId()}
                   },'*');
                 }
                 else if (event.data.module === 'comfirmPrompt') {