+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.52.0]
+* Silent Monitoring - Re-add feature flag
+* <CXV1-12475> - add abiliity to open outbound identifier page via url
+
 ## [4.51.0]
 * CXV1-12492 - TelStrat Artifact recording support